Fixed code maintenability.
