+++ app/react/Charts/components/StackedDualBarChart.js
@@ -2,18 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { ResponsiveContainer, BarChart, XAxis, YAxis, CartesianGrid, Bar, Tooltip, Legend } from 'recharts';
 
+import { formatPayload } from '../utils/arrayUtils';
+
 import ExtendedTooltip from './ExtendedTooltip';
 import ColoredBar from './ColoredBar';
-import colorScheme from '../utils/colorScheme';
 
 const StackedDualBarChart = (props) => {
   const { data, chartLabel } = props;
-  const legendPayload = data.map((item, index) => ({
-    value: item.name,
-    type: 'rect',
-    color: colorScheme[index % colorScheme.length],
-    formatter: () => <span style={{ color: '#333' }}>{item.name}</span>
-  }));
 
   return (
     <ResponsiveContainer height={320}>
@@ -24,7 +19,7 @@ const StackedDualBarChart = (props) => {
         <Tooltip content={<ExtendedTooltip parentData={data} chartLabel={chartLabel} />}/>
         <Bar dataKey="setAValue" fill="#D24040" shape={<ColoredBar />} stackId="unique" />
         <Bar dataKey="setBValue" fill="#D24040" shape={<ColoredBar color="light" />} stackId="unique" />
-        <Legend payload={legendPayload} />
+        <Legend payload={formatPayload(data)} />
       </BarChart>
     </ResponsiveContainer>
   );
