+++ app/react/Charts/index.js
@@ -2,12 +2,12 @@ import ColoredBar from './components/ColoredBar';
 import ExtendedTooltip from './components/ExtendedTooltip';
 import StackedDualBarChart from './components/StackedDualBarChart';
 import colorScheme from './utils/colorScheme';
-import sortUtils from './utils/sortUtils';
+import arrayUtils from './utils/arrayUtils';
 
 export default {
   ColoredBar,
   ExtendedTooltip,
   StackedDualBarChart,
   colorScheme,
-  sortUtils
+  arrayUtils
 };
